normalize issuer as a sanity check
Inexperienced devs/users may not know whether to include a final
slash (/) at the end of their issuer field in their metadata endpoint.
Also, it seems some OIDC/OAuth modules may require there to NOT
be a final slash (/) at the end of the issuer.

This change is intended to make it less error prone overall.